libs/foo/pyproject.toml
[tool.poetry]
name = "Foo"
version = "1.0.1"
description = "The Foo Python library"
authors = [
    "Charles <charles@org.com>",
]

# Ref: https://github.com/python-poetry/poetry/issues/167#issuecomment-530895533
packages = [ { include = "org", from = "src" } ]

[tool.poetry.dependencies]
python = "^3.11"

pillow = "^10.1"

[tool.poetry.group.dev.dependencies]
# testing
pytest = "^7.4"
# linting
black = "^23.12"
mypy = "^1.8"
ruff = "^0.1"
# typing
types-pillow = "^10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
